requirements.txt
pytest
pyaml
docopts
subprocess32
six
